FEAT: REMOVED GOOGLE SHEETS
LETS GOOO, NO MORE STUPID GOOGLE SHEET API ERRORS!!!! 🥳🥳

ALL BACKUPS ON FIREBASE 🎉🎉
