+++ src/attending-server/base-attending-server.ts
@@ -16,7 +16,6 @@ import {
 } from 'discord.js';
 import { environment } from '../environment/environment-manager.js';
 import { ServerExtension } from '../extensions/extension-interface.js';
-import { GoogleSheetServerExtension } from '../extensions/google-sheet-logging/google-sheet-server-extension.js';
 import { CalendarServerExtension } from '../extensions/session-calendar/calendar-server-extension.js';
 import { LOGGER } from '../global-states.js';
 import { AutoClearTimeout, HelpQueue } from '../help-queue/help-queue.js';
@@ -291,7 +290,6 @@ class AttendingServer {
         const serverExtensions: readonly ServerExtension[] = environment.disableExtensions
             ? []
             : await Promise.all([
-                  GoogleSheetServerExtension.load(guild),
                   CalendarServerExtension.load(guild),
                   new HelperActivityTrackingExtension(guild)
               ]);
