+++ src/extensions/helper-activity-tracking/constants/interaction-names.ts
@@ -4,7 +4,11 @@ enum ActivityTrackingCommandNames {
     dump_tracking_data = 'dump_tracking_data'
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-type A = EnsureCorrectEnum<typeof ActivityTrackingCommandNames>;
+enum ActivityTrackingButtonNames {
+    UpdateTrackingStatus = 'UpdateTrackingStatus',
+}
 
-export { ActivityTrackingCommandNames };
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+type A = EnsureCorrectEnum<typeof ActivityTrackingCommandNames> & EnsureCorrectEnum<typeof ActivityTrackingButtonNames>;
+
+export { ActivityTrackingCommandNames, ActivityTrackingButtonNames};
