+++ src/extensions/helper-activity-tracking/constants/success-messages.ts
@@ -1,19 +1,14 @@
 import { EmbedColor, SimpleEmbed } from '../../../utils/embed-helper.js';
 import { YabobEmbed } from '../../../utils/type-aliases.js';
 
-const GoogleSheetSuccessMessages = {
-    updatedGoogleSheet: (newSheetTitle: string): YabobEmbed =>
-        SimpleEmbed(
-            `Successfully changed to this new google sheet: ${newSheetTitle}`,
-            EmbedColor.Success
-        ),
+const ActivityTrackingSuccessMessages = {
     updatedSheetTracking: (newTrackingStatus: boolean): YabobEmbed =>
         SimpleEmbed(
             `Successfully ${
                 newTrackingStatus ? 'enabled' : 'disabled'
-            } google sheet tracking`,
+            } activity tracking`,
             EmbedColor.Success
         )
 } as const;
 
-export { GoogleSheetSuccessMessages };
+export { ActivityTrackingSuccessMessages };
